@@ -1,22 +1,31 @@
 {
-  "name": "javascript-development-environment",
+  "name": "js-dev-env",
   "version": "1.0.0",
   "description": "JavaScript dev environment boilerplate by cory house",
   "scripts": {
+    "prestart": "babel-node buildScripts/startMessage.js",
+    "start": "npm-run-all --parallel security-check open:src",
+    "open:src": "babel-node buildScripts/srcServer.js",
+    "security-check": "nsp check",
+    "localtunnel": "lt --port 3000",
+    "share": "npm-run-all --parallel open:src localtunnel"
   },
   "author": "Cory House",
   "license": "MIT",
   "dependencies": {
-    "whatwg-fetch": "1.0.0"
+    "webpack": "^3.8.1",
+    "webpack-dev-middleware": "^1.12.0",
+    "webpack-hot-middleware": "^2.20.0",
+    "whatwg-fetch": "^2.0.3"
   },
   "devDependencies": {
-    "babel-cli": "6.16.0",
-    "babel-core": "6.17.0",
+    "babel-cli": "6.26.0",
+    "babel-core": "6.26.0",
     "babel-loader": "6.2.5",
     "babel-preset-latest": "6.16.0",
     "babel-register": "6.16.3",
-    "chai": "3.5.0",
-    "chalk": "1.1.3",
+    "chai": "^4",
+    "chalk": "^2",
     "cheerio": "0.22.0",
     "compression": "1.6.2",
     "cross-env": "3.1.3",
@@ -39,9 +48,9 @@
     "open": "0.0.5",
     "rimraf": "2.5.4",
     "style-loader": "0.13.1",
-    "webpack": "1.13.2",
-    "webpack-dev-middleware": "1.8.4",
-    "webpack-hot-middleware": "2.13.0",
+    "webpack": "3.8.1",
+    "webpack-dev-middleware": "1.12.0",
+    "webpack-hot-middleware": "2.20.0",
     "webpack-md5-hash": "0.0.5"
   }
 }
